Edge/data_injector/requirements.txt
setuptools
requests
paho-mqtt
